+++ app/Models/Trailers.php
@@ -2,7 +2,6 @@
 
 namespace App\Models;
 
-use GeneaLabs\LaravelModelCaching\Traits\Cachable;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Notifications\Notifiable;
@@ -14,8 +13,6 @@ class Trailers extends Model implements \OwenIt\Auditing\Contracts\Auditable
     use HasFactory;
     use HasFactory, Notifiable, HasRoles;
     use Auditable;
-    use Cachable;
-
     /**
      * The attributes that are mass assignable.
      *