+++ resources/assets/core/js/components/dialer.js
@@ -95,7 +95,7 @@ var KTDialer = function(element, options) {
             _decrease();
         });
 
-        KTUtil.addEvent(the.inputElement, 'change', function(e) {
+        KTUtil.addEvent(the.inputElement, 'input', function(e) {
             e.preventDefault();
 
             _setValue();
@@ -121,6 +121,7 @@ var KTDialer = function(element, options) {
         KTEventHandler.trigger(the.element, 'kt.dialer.decrease', the);
 
         the.inputElement.value = the.value - the.options.step;        
+        console.log('dec:' + the.inputElement.value);
         _setValue();
 
         // Trigger "before.dialer" event
@@ -134,24 +135,48 @@ var KTDialer = function(element, options) {
         // Trigger "after.dialer" event
         KTEventHandler.trigger(the.element, 'kt.dialer.change', the);
 
-        the.value = parseFloat(the.inputElement.value.replace(/[^\d.]/g, '')); 
+        the.value = _parse(the.inputElement.value); 
+
+        console.log('min:' + the.options.min);
         
-        if (the.value < the.options.min) {
+        if (the.options.min !== null && the.value < the.options.min) {
             the.value = the.options.min;
+            console.log('reset: min');
         }
 
-        if (the.value > the.options.max) {
+        if (the.options.max !== null && the.value > the.options.max) {
             the.value = the.options.max;
         }
 
         the.inputElement.value = _format(the.value);
 
+        // Trigger input change event
+        the.inputElement.dispatchEvent(new Event('change'));
+
         // Trigger "after.dialer" event
         KTEventHandler.trigger(the.element, 'kt.dialer.changed', the);
     }
 
+    var _parse = function(val) {
+        val = val
+            .replace(/[^0-9.-]/g, '')       // remove chars except number, hyphen, point. 
+            .replace(/(\..*)\./g, '$1')     // remove multiple points.
+            .replace(/(?!^)-/g, '')         // remove middle hyphen.
+            .replace(/^0+(\d)/gm, '$1');    // remove multiple leading zeros. <-- I added this.
+
+        val = parseFloat(val);
+
+        if (isNaN(val)) {
+            val = 0;
+        } 
+
+        return val;
+    }
+
     // Format
     var _format = function(val){
+        console.log('val:' + val);
+
         return the.options.prefix + parseFloat(val).toFixed(the.options.decimals) + the.options.suffix;              
     }