+++ resources/views/layout/demo8/_content.blade.php
@@ -1,5 +1,7 @@
 <!--begin::Container-->
-<div class="container-fluid d-flex flex-stack">
+<div class="d-flex flex-column flex-column-fluid container-fluid">
+<div class="content flex-column-fluid" id="kt_content">
     {{ $slot }}
 </div>
 <!--end::Container-->
+</div>