+++ resources/views/ordertypes/index.blade.php
@@ -9,7 +9,35 @@
     </script>
 @endforeach
 <script>
-    $(function(){$('#order_type_table').DataTable({processing:!0,serverSide:!0,searching:!1,"order":[[0,"desc"]],search:{return:!0},pageLength:7,ajax:'{!! route('ordertype.list') !!}',columns:[{data:'status',name:'status'},{data:'order_type_id',name:'order_type_id'},{data:'description',name:'description'},{data:'Actions',name:'Actions',orderable:!1,serachable:!1,sClass:'text-center'},]})})
+    $(function(){$('#order_type_table').DataTable(
+        {
+        processing:!0,
+            serverSide:!0,
+            searching:!1,
+            "order":[[0,"desc"]],
+            search:{return:!0},
+            pageLength:7,
+            ajax:'{!! route('ordertype.list') !!}',
+            columns:[
+                {
+                    data: "status",
+                    render: function (data, type, row, meta) {
+                        if (type === 'display') {
+                            if (data === 'Active') {
+                                data = '<span class="badge badge-light-success">Active</span>';
+                            } else if (data === 'Inactive') {
+                                data = '<span class="badge badge-light-danger">Inactive</span>';
+                            }
+                            return data;
+                        }
+                    }
+                },
+                {data:'order_type_id', name:'order_type_id'},
+                {data:'description', name:'description'},
+                {data:'Actions', name:'Actions', orderable:!1, serachable:!1, sClass:'text-center'},
+        ]
+    })
+    })
 </script>
 <script type="text/javascript">
     var CSRF_TOKEN = $('meta[name="csrf-token"]').attr("content");