+++ resources/views/tractors/index.blade.php
@@ -24,7 +24,32 @@
 
 <!-- begin::Script to Produce DataTable for Tractors -->
 <script>
-    $(function(){$('#tractor_table').DataTable({processing:!0,serverSide:!0,searching:!1,"order":[[0,"desc"]],search:{return:!0},pageLength:7,ajax:'{!! route('tractors.list') !!}',columns:[{data:'status',name:'status'},{data:'tractor_id',name:'tractor_id'},{data:'year',name:'year'},{data:'make',name:'make'},{data:'model',name:'model'},{data:'owned_by',name:'owned_by'},{data:'last_inspection',name:'last_inspection'},{data:'Actions',name:'Actions',orderable:!1,serachable:!1,sClass:'text-center'},]})})
+    $(function(){$('#tractor_table').DataTable({processing:!0,serverSide:!0,searching:!1,"order":[[0,"desc"]],
+        search:{return:!0},
+        pageLength:7,
+        ajax:'{!! route('tractors.list') !!}',
+        columns:[
+        {
+            data: "status",
+            render: function (data, type, row, meta) {
+                if (type === 'display') {
+                    if (data === 'Active') {
+                        data = '<span class="badge badge-light-success">Active</span>';
+                    } else if (data === 'Inactive') {
+                        data = '<span class="badge badge-light-danger">Inactive</span>';
+                    }
+                    return data;
+                }
+            }
+        },
+            {data:'tractor_id',name:'tractor_id'},
+            {data:'year',name:'year'},
+            {data:'make',name:'make'},
+            {data:'model',name:'model'},
+            {data:'owned_by',name:'owned_by'},
+            {data:'last_inspection',name:'last_inspection'},
+            {data:'Actions',name:'Actions',orderable:!1,serachable:!1,sClass:'text-center'},
+        ]})})
 </script>
 <!-- end::Script to Produce DataTable for Tractors -->