+++ resources/views/trailers/index.blade.php
@@ -45,29 +45,29 @@
             order: [[0, "desc"]],
             search: {return: !0},
             pageLength: 7,
-            ajax: '{!! route('trailers.list') !!}',
-            columns: [
-                {
-                    data: "status",
-                    render: function (data, type, row, meta) {
-                        if (type === 'display') {
-                            if (data === 'Active') {
-                                data = '<span class="badge badge-light-success">Active</span>';
-                            } else if (data === 'Inactive') {
-                                data = '<span class="badge badge-light-danger">Inactive</span>';
+                ajax: '{!! route('trailers.list') !!}',
+                columns: [
+                    {
+                        data: "status",
+                        render: function (data, type, row, meta) {
+                            if (type === 'display') {
+                                if (data === 'Active') {
+                                    data = '<span class="badge badge-light-success">Active</span>';
+                                } else if (data === 'Inactive') {
+                                    data = '<span class="badge badge-light-danger">Inactive</span>';
+                                }
+                                return data;
                             }
-                            return data;
                         }
-                    }
-                },
-                {data: "trailer_id", name: "trailer_id"},
-                {data: "year", name: "year"},
-                {data: "make", name: "make"},
-                {data: "model", name: "model"},
-                {data: "owned_by", name: "owned_by"},
-                {data: "last_inspection", name: "last_inspection"},
-                {data: "Actions", name: "Actions", orderable: !1, serachable: !1, sClass: "text-center"},
-            ],
+                    },
+                    {data: "trailer_id", name: "trailer_id"},
+                    {data: "year", name: "year"},
+                    {data: "make", name: "make"},
+                    {data: "model", name: "model"},
+                    {data: "owned_by", name: "owned_by"},
+                    {data: "last_inspection", name: "last_inspection"},
+                    {data: "Actions", name: "Actions", orderable: !1, serachable: !1, sClass: "text-center"},
+                ],
         });
     });
 </script>